Erase warnings on 64bit build

+++ lua-5.3.1/src/lstrlib.c
@@ -836,7 +836,8 @@ static int num2straux (char *buff, lua_Number x) {
   else if (x == 0) {  /* can be -0... */
     sprintf(buff, LUA_NUMBER_FMT, x);
     strcat(buff, "x0p+0");  /* reuses '0/-0' from 'sprintf'... */
-    return strlen(buff);
+	// warning C4267: 'return': conversion from 'size_t' to 'int', possible loss of data
+    return (int)strlen(buff);
   }
   else {
     int e;
